"cannot access the file because it is being used"
With Maven / Main.kt and Windows, closing modified jar files resulted
in an exception from ZipFileSystem because the jar was opened by a
classloader.

Resolve this by making a temporary copies of the files before
processing, so the original copy can be used by the classloader and the
copy can be modified and then used to replace the original.